Prototype implementation of gallery image creation
This is an example of how to request the creation of a gallery. Currently, the request is hardcoded. We will need to pass data from the gallery page to the download page to create this request dynamically...

The image util class should be in its final stages. There are still a few open design questions for the overlay function, though. Namely, how do we want images to be spaced?