Perform blockchain queries as a background task
Users are redirected to a loading screen while the blockchain is being queried for NFTs. The loading screen polls a background task queue until the query is complete. Consult the README for updated instructions on how to run the server.

NOTE: SHOULD NOT BE DEPLOYED UNTIL REDISTOGO HAS BEEN ADDED TO HEROKU.